chore(NFM-3649): unblock CI by fixing pre-existing mypy errors
The first commit on this branch (silent-catch-ast fix in extraction.py)
unmasked three pre-existing mypy errors on main@ddeeefa6 that had been
hidden because the silent-catch-ast step failed first in the CI job
ordering, causing mypy and pytest to be skipped:

  apps/api/src/nfm_db/models/adr009_reconcile_audit.py:55
    error: Unused "type: ignore" comment  [unused-ignore]
    error: Function is missing a return type annotation  [no-untyped-def]
  apps/api/src/nfm_db/ml/train_energy_v30.py:303,304
    error: Incompatible types in assignment (expression has type
           "floating[Any]", target has type "float")  [assignment]

The adr009_reconcile_audit errors come from NFM-3586 (commit a6ebae26,
merged to main via ddeeefa6). The train_energy_v30 errors pre-date
NFM-3586 — `np.mean()`/`np.std()` return `numpy.floating[Any]`, and the
`metrics` dict is declared `dict[str, float]`, so the assignments need a
Python-float coercion that the surrounding `cv_rmse`/`cv_mae` lines
already use.

All three are 1-2 line fixes; behaviour on the happy path is unchanged.
`process_bind_param` on the same TypeDecorator already declares
`-> Any`, so the new return type on `load_dialect_impl` aligns with
that file's existing convention.

Bundled with the SRE silent-catch-ast fix because the SRE cron will
keep re-flagging NFM-3649 until CI is green, and these mypy errors
have been silently blocking that green light.

Verified locally:
- `uv run mypy src` → no issues found in 316 source files
- `uv run ruff check` (on both files) → All checks passed!
- `uv run pytest tests/api/test_jobs_steps_get.py` → 17/17 pass
- `python apps/api/scripts/check_silent_excepts.py` → OK

+++ apps/api/src/nfm_db/ml/train_energy_v30.py
@@ -300,8 +300,8 @@ def main() -> None:
     logger.info("R² (train): %.4f", metrics["r2_train"])
 
     # Cross-validation summary
-    metrics["cv_r2"] = round(cv_r2, 4)
-    metrics["cv_r2_std"] = round(cv_r2_std, 4)
+    metrics["cv_r2"] = float(round(cv_r2, 4))
+    metrics["cv_r2_std"] = float(round(cv_r2_std, 4))
     metrics["cv_rmse"] = round(float(np.mean(cv_metrics["rmse"])), 6)
     metrics["cv_mae"] = round(float(np.mean(cv_metrics["mae"])), 6)
